NetworkManager added
