+++ UIKit-RandomCinema/Model/Cinema.swift
@@ -13,7 +13,6 @@ struct Cinema: Codable {
 
 struct Doc: Codable {
     let poster: Poster
-    let rating: Rating
     let name: String
     let description: String
     let year: Int
@@ -25,6 +24,3 @@ struct Poster: Codable {
     let url: String
 }
 
-struct Rating: Codable {
-    let imdb: Double
-}
